Cross-compile testSuite/../javalib/math.
