+++ test-suite/js/src/test/scala/org/scalajs/testsuite/utils/Platform.scala
@@ -1,0 +1,8 @@
+package org.scalajs.testsuite.utils
+
+object Platform {
+  /** Returns `true` if and only if the code is executing on a JVM.
+   *  Note: Returns `false` when executing on any JS VM.
+   */
+  final val executingInJVM = false
+}
