+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/javalib/math/BigDecimalTest.scala
