+++ test-suite/shared/src/test/scala/org/scalajs/testsuite/javalib/math/BigIntegerConvertTest.scala
@@ -9,7 +9,6 @@ package org.scalajs.testsuite.javalib.math
 
 import java.math.BigInteger
 
-import org.scalajs.jasminetest.JasmineTest
 import org.junit.{Test, Ignore}
 import org.junit.Assert._
 
@@ -548,19 +547,5 @@ class BigIntegerConvertTest {
        assertEquals(rBytes(i), resBytes(i))
     }
     assertEquals(0, aNumber.signum())
-  }
-}
-
-object BigIntegerConvertJSTest extends JasmineTest {
-  describe("BigIntegerConvertTest") {
-    // To test that it works with strict floats, do:
-    //   > set scalaJSSemantics in testSuite ~= { _.withStrictFloats(true) }
-    when("strict-floats").
-    it("testDoubleValueZero") {
-      val a = "0"
-      val result = 0.0
-      val aNumber = new BigInteger(a).doubleValue()
-      expect(aNumber).toEqual(result)
-    }
   }
 }
